Refactored sleep for sleep modes
