+++ Application/Inc/SYS/sleep/sleep_config.h
@@ -1,12 +1,12 @@
 /*
- * sys_main.h
+ * sleep_config.h
  *
- *  Created on: Mar 11, 2025
+ *  Created on: Aug 2, 2025
  *      Author: ericv
  */
 
-#ifndef SYS_SYS_MAIN_H_
-#define SYS_SYS_MAIN_H_
+#ifndef SLEEP_SLEEP_CONFIG_H_
+#define SLEEP_SLEEP_CONFIG_H_
 
 #ifdef __cplusplus
 extern "C" {
@@ -22,7 +22,11 @@ extern "C" {
 
 /* Exported types ------------------------------------------------------------*/
 
-
+typedef enum {
+  SLEEP_NONE,
+  SLEEP_LIGHT,
+  SLEEP_DEEP
+} SleepStates_t;
 
 /* Exported constants --------------------------------------------------------*/
 
@@ -30,13 +34,11 @@ extern "C" {
 
 /* Exported macro ------------------------------------------------------------*/
 
-#define SLEEP_REQUEST_LIGHT   (1 << 0)
-#define SLEEP_REQUEST_DEEP    (1 << 1)
-#define SLEEP_WAKEUP_MESS     (1 << 2)
+
 
 /* Exported functions prototypes ---------------------------------------------*/
 
-void SYS_StartTask(void* argument);
+
 
 /* Private defines -----------------------------------------------------------*/
 
@@ -44,4 +46,4 @@ void SYS_StartTask(void* argument);
 }
 #endif
 
-#endif /* SYS_SYS_MAIN_H_ */
+#endif /* SLEEP_SLEEP_CONFIG_H_ */
