+++ Application/Inc/SYS/sleep/sleep_deep.h
@@ -1,12 +1,12 @@
 /*
- * sys_sleep.h
+ * sleep_deep.h
  *
- *  Created on: Jul 27, 2025
+ *  Created on: Aug 2, 2025
  *      Author: ericv
  */
 
-#ifndef SYS_SYS_SLEEP_H_
-#define SYS_SYS_SLEEP_H_
+#ifndef SLEEP_SLEEP_DEEP_H_
+#define SLEEP_SLEEP_DEEP_H_
 
 #ifdef __cplusplus
 extern "C" {
@@ -14,7 +14,7 @@ extern "C" {
 
 /* Includes ------------------------------------------------------------------*/
 #include "stm32h7xx_hal.h"
-#include <stdbool.h>
+
 
 /* Private includes ----------------------------------------------------------*/
 
@@ -30,14 +30,11 @@ extern "C" {
 
 /* Exported macro ------------------------------------------------------------*/
 
-#define SLEEP_REQUEST_LIGHT   (1 << 0)
-#define SLEEP_REQUEST_DEEP    (1 << 1)
-#define SLEEP_WAKEUP_MESS     (1 << 2)
+
 
 /* Exported functions prototypes ---------------------------------------------*/
 
-bool Sleep_Init();
-void Sleep_Check();
+void SleepDeep_Enter();
 
 /* Private defines -----------------------------------------------------------*/
 
@@ -45,4 +42,4 @@ void Sleep_Check();
 }
 #endif
 
-#endif /* SYS_SYS_SLEEP_H_ */
+#endif /* SLEEP_SLEEP_DEEP_H_ */
