+++ Application/Src/SYS/sleep/sleep_config.c
@@ -1,0 +1,36 @@
+/*
+ * sleep_config.c
+ *
+ *  Created on: Aug 2, 2025
+ *      Author: ericv
+ */
+
+/* Private includes ----------------------------------------------------------*/
+
+#include "sleep/sleep_config.h"
+
+/* Private typedef -----------------------------------------------------------*/
+
+
+
+/* Private define ------------------------------------------------------------*/
+
+
+
+/* Private macro -------------------------------------------------------------*/
+
+
+
+/* Private variables ---------------------------------------------------------*/
+
+
+
+/* Private function prototypes -----------------------------------------------*/
+
+
+
+/* Exported function definitions ---------------------------------------------*/
+
+
+
+/* Private function definitions ----------------------------------------------*/
