+++ Application/Src/SYS/sleep/wakeup_tones.c
@@ -1,0 +1,36 @@
+/*
+ * wakeup_tones.c
+ *
+ *  Created on: Aug 2, 2025
+ *      Author: ericv
+ */
+
+/* Private includes ----------------------------------------------------------*/
+
+#include "sleep/wakeup_tones.h"
+
+/* Private typedef -----------------------------------------------------------*/
+
+
+
+/* Private define ------------------------------------------------------------*/
+
+
+
+/* Private macro -------------------------------------------------------------*/
+
+
+
+/* Private variables ---------------------------------------------------------*/
+
+
+
+/* Private function prototypes -----------------------------------------------*/
+
+
+
+/* Exported function definitions ---------------------------------------------*/
+
+
+
+/* Private function definitions ----------------------------------------------*/
