+++ Core/Inc/main.h
@@ -62,7 +62,6 @@ void HAL_TIM_MspPostInit(TIM_HandleTypeDef *htim);
 void Error_Handler(void);
 
 /* USER CODE BEGIN EFP */
-void reinitializePeripherals(void);
 /* USER CODE END EFP */
 
 /* Private defines -----------------------------------------------------------*/
